Updated version numbers to 0.9.7-pre.
git-svn-id: svn://svn.code.sf.net/p/unfs3/code/unfs3.svn/trunk@50 fb7659fb-8760-4330-86f7-fdbeb49e81a3

+++ unfs3.spec
@@ -1,12 +1,13 @@
+%define version 0.9.7
 Summary: UNFS3 user-space NFSv3 server
 Group: Applications/System
 Name: unfs3
-Version: 0.9.6
+Version: %{version}
 Release: 1
 Copyright: BSD
 Packager: Pascal Schmidt <der.eremit@email.de>
 Vendor: none
-Source: unfs3-0.9.6.tar.gz
+Source: unfs3-%{version}.tar.gz
 Buildroot: /tmp/unfs3
 
 %prep
@@ -17,10 +18,11 @@ Buildroot: /tmp/unfs3
 make
 
 %install
+[ -n "$RPM_BUILD_ROOT" ] && rm -rf "$RPM_BUILD_ROOT"
 make DESTDIR="$RPM_BUILD_ROOT" install
 
 %clean
-rm -rf "$RPM_BUILD_ROOT"
+[ -n "$RPM_BUILD_ROOT" ] && rm -rf "$RPM_BUILD_ROOT"
 
 %description
 UNFS3 is a user-space implementation of the NFS (Network File System)
@@ -28,7 +30,7 @@ version 3 server specification. It provides a daemon that supports both
 the MOUNT and NFS protocol.
 
 %files
-%attr (-, root, root) %doc README LICENSE NEWS
+%attr (  -, root, root) %doc README LICENSE NEWS
 %attr (755, root, root) /usr/sbin/unfsd
 %attr (644, root, root) %doc /usr/share/man/man7/tags.7.gz
 %attr (644, root, root) %doc /usr/share/man/man8/unfsd.8.gz
